requirements.txt
python-dotenv==1.2.2
clickhouse-driver==0.2.10
psycopg==3.3.3
websockets==16.0
pyyaml==6.0.3
boto3==1.42.82